% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{brier}
\alias{brier}
\title{Brier Score}
\usage{
brier(act, pred)
}
\arguments{
\item{act}{actual result(s)}

\item{pred}{predicted result(s)}
}
\value{
the brier score for this data or series of data
}
\description{
brier() returns the Brier score obtained from a given prediction and known result
}
