% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{db.create_table}
\alias{db.create_table}
\title{Create Table in DB}
\usage{
db.create_table(con, what = c("roster", "shifts", "pbp"))
}
\arguments{
\item{con}{A connection to the database (see \code{\link{db.get_db_connection}()}).}

\item{what}{Which of roster, shifts and/or pbp table to create in the database .}
}
\description{
Create the table in the database, if it doesn't already exist
}
