% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{db.get_db_connection}
\alias{db.get_db_connection}
\title{Get Database Connection}
\usage{
db.get_db_connection(user = "hockey", password = "analytic",
  dbname = "hockeystats", dbtype = "PostgreSQL")
}
\arguments{
\item{user}{the username to connect to the database}

\item{password}{the password to connect to the database}

\item{dbname}{the database's name}

\item{dbtype}{the database type. Only tested with PostgreSQL databases}
}
\value{
con = a database connection for use
}
\description{
A convenience function for connecting to the database using RPostgreeSQL
}
