% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{dcapply}
\alias{dcapply}
\title{Do Call Apply}
\usage{
dcapply(x, fun, combine, cores, ...)
}
\arguments{
\item{x}{the data to split and apply a function}

\item{fun}{the function to apply to the data}

\item{combine}{the method to recombine the data}

\item{cores}{the number of cores to use}

\item{...}{additional parameters to lapply on each chunk}
}
\value{
The data with the function applied
}
\description{
dcapply() uses do.call() to merge the products of an applied function according to specifications. The function will be applied in parallel if cores >= 1
}
