% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{ds.clean_nums}
\alias{ds.clean_nums}
\title{Clean Nums
clean_nums() returns a list of player number identifiers for a given event description}
\usage{
ds.clean_nums(x)
}
\arguments{
\item{x}{event description(s)}
}
\value{
list
}
\description{
Clean Nums
clean_nums() returns a list of player number identifiers for a given event description
}
\keyword{internal}
