% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{ds.compile_games}
\alias{ds.compile_games}
\title{Compile Games}
\usage{
ds.compile_games(games, season, pause = 1, try_tolerance = 3,
  agents = hockeyR::ds.get_user_agents()[sample(1:4, size = 1)])
}
\arguments{
\item{games}{the specific games to compile}

\item{season}{the season to compile (games are in this chosen season)}

\item{pause}{delay between compilations}

\item{try_tolerance}{the number of tries before skipping the game}

\item{agents}{user agents, selected from a set.}
}
\value{
list object containing c([[1]] = PBP, [[2]] = Roster, [[3]] = Shifts) is returned
}
\description{
compile_games() collects, parses and compiles all game data corresponding to a given vector of game IDs and season.
}
