% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{ds.find_goalie}
\alias{ds.find_goalie}
\title{Is On Goalie}
\usage{
ds.find_goalie(players, roster)
}
\arguments{
\item{players}{a vector of players}

\item{roster}{roster information}
}
\value{
a vector of goalies, if any, else NA
}
\description{
find_goalie() returns a vector containing all goaltenders in a given player vector
}
