% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{ds.fix_names}
\alias{ds.fix_names}
\title{Fix Names}
\usage{
ds.fix_names(name_vect)
}
\arguments{
\item{name_vect}{a vector of player names}
}
\value{
a corrected vector of player names
}
\description{
fix_names() returns a vector of player names corrected for multiple spelling variants
}
