% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{ds.get_coordinates}
\alias{ds.get_coordinates}
\title{Get Coordinates}
\usage{
ds.get_coordinates(season, game_id, source, date, away_team,
  try_tolerance = 3, agents = hockeyR::ds.get_user_agents()[sample(1:4, size
  = 1)])
}
\arguments{
\item{season}{a season}

\item{game_id}{a single game id}

\item{source}{espn or nhl}

\item{date}{the date of the game (for espn)}

\item{away_team}{the away team}

\item{try_tolerance}{the number of tries before moving on.}

\item{agents}{a useragent, selected from the list of useragents.}
}
\value{
list object
}
\description{
get_coordinates() imports the event coordinates corresponding to a given year and game ID and returns a list object.
}
