% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{ds.get_highlights}
\alias{ds.get_highlights}
\title{Get Highlights}
\usage{
ds.get_highlights(season, game_id, try_tolerance = 3,
  agents = hockeyR::ds.get_user_agents()[sample(1:4, size = 1)])
}
\arguments{
\item{season}{a season}

\item{game_id}{a single game id}

\item{try_tolerance}{the number of tries before moving on.}

\item{agents}{a useragent, selected from the list of useragents.}
}
\value{
a json list object
}
\description{
get_highlights() imports the highlights page corresponding to a given year and game ID and returns a JSON list object.
}
\keyword{internal}
