% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{ds.get_pbp}
\alias{ds.get_pbp}
\title{Get PBP}
\usage{
ds.get_pbp(season, game_id, try_tolerance = 3,
  agents = ds.user_agents[sample(1:4, size = 1)])
}
\arguments{
\item{season}{a season}

\item{game_id}{a single game id}

\item{try_tolerance}{the number of tries before moving on.}

\item{agents}{a useragent, selected from the list of useragents.}
}
\value{
a list object
}
\description{
get_pbp() imports the PBP page corresponding to a given year and game ID and returns a list object.
}
