% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{ds.get_player_profile}
\alias{ds.get_player_profile}
\title{Get Player Profile}
\usage{
ds.get_player_profile(player_id, try_tolerance = 3,
  agents = hockeyR::ds.get_user_agents()[sample(1:4, size = 1)])
}
\arguments{
\item{player_id}{the player id}

\item{try_tolerance}{the number of tries before moving on.}

\item{agents}{a useragent, selected from the list of useragents.}
}
\value{
json list objet
}
\description{
get_player_profile() imports the player profile page corresponding to a given player ID and returns a JSON list object.
}
\keyword{internal}
