% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{ds.get_roster}
\alias{ds.get_roster}
\title{Get Roster}
\usage{
ds.get_roster(season, game_id, try_tolerance = 3,
  agents = hockeyR::ds.get_user_agents()[sample(1:4, size = 1)])
}
\arguments{
\item{season}{a season}

\item{game_id}{a single game id}

\item{try_tolerance}{the number of tries before moving on.}

\item{agents}{a useragent, selected from the list of useragents.}
}
\value{
character vector
}
\description{
get_roster() imports the Roster page corresponding to a given year and game ID and returns a character vector
}
