% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{ds.get_schedule}
\alias{ds.get_schedule}
\title{Get Schedule}
\usage{
ds.get_schedule(start, end, try_tolerance = 3,
  agents = hockeyR::ds.get_user_agents()[sample(1:4, size = 1)])
}
\arguments{
\item{start}{Start Date}

\item{end}{End Date}

\item{try_tolerance}{the number of tries before moving on.}

\item{agents}{a useragent, selected from the list of useragents.}
}
\value{
a json list object
}
\description{
get_schedule() imports the schedule page corresponding to a given date range and returns a JSON list object.
}
