% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{ds.get_shifts}
\alias{ds.get_shifts}
\title{Get Shifts}
\usage{
ds.get_shifts(season, game_id, venue, source, try_tolerance = 3,
  agents = hockeyR::ds.get_user_agents()[sample(1:4, size = 1)])
}
\arguments{
\item{season}{a season}

\item{game_id}{a single game id}

\item{venue}{home or away}

\item{source}{htm or json}

\item{try_tolerance}{the number of tries before moving on.}

\item{agents}{a useragent, selected from the list of useragents.}
}
\value{
a list object
}
\description{
get_shifts() imports the shift report page corresponding to a given year, game ID and venue and returns a list object.
}
