% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{ds.get_team_profile}
\alias{ds.get_team_profile}
\title{Get Team Profile}
\usage{
ds.get_team_profile(team_id, try_tolerance = 3,
  agents = hockeyR::ds.get_user_agents()[sample(1:4, size = 1)])
}
\arguments{
\item{team_id}{the team id to get}

\item{try_tolerance}{the number of tries before moving on.}

\item{agents}{a useragent, selected from the list of useragents.}
}
\value{
json list
}
\description{
get_team_profile() imports the team profile page corresponding to a given team ID and returns a JSON list object.
}
\keyword{internal}
