% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{ds.is_on}
\alias{ds.is_on}
\title{Is On}
\usage{
ds.is_on(player, pbp, venue)
}
\arguments{
\item{player}{a vector of players}

\item{pbp}{a pbp object}

\item{venue}{home or away}
}
\value{
numeric vector of 1 or 0 indicating player on ice
}
\description{
is_on() returns a numeric vector indicating 1 if a given player is on ice during the event corresponding to the row index in the given PBP object
}
