% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{ds.merge_season}
\alias{ds.merge_season}
\title{Merge Season Files}
\usage{
ds.merge_season(season_dir, what = c("roster", "shifts", "pbp"))
}
\arguments{
\item{season_dir}{The directory of the season to merge}

\item{what}{What to merge. One or more of `roster`, `shifts`, or `pbp`.}
}
\description{
Merge season files in a directory, looking for .roster, .shifts or .pbp files to merge. Will leak readr::read_delim and dplyr::bind_rows errors.
}
