% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{ds.ms_from_seconds}
\alias{ds.ms_from_seconds}
\title{MM:SS from seconds}
\usage{
ds.ms_from_seconds(seconds)
}
\arguments{
\item{seconds}{Number of seconds}
}
\value{
MM:SS of play time in that period (returns from 00:00 to 20:00, 1201 seconds = 00:01)
}
\description{
MM:SS from seconds
}
\keyword{internal}
