% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{ds.parse_event}
\alias{ds.parse_event}
\title{Parse PBP Event}
\usage{
ds.parse_event(x)
}
\arguments{
\item{x}{pbp json}
}
\value{
data frame
}
\description{
parse_event() parses a single event from the PBP JSON object and returns a data frame
}
