% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{ds.parse_game}
\alias{ds.parse_game}
\title{Parse Game}
\usage{
ds.parse_game(x)
}
\arguments{
\item{x}{the schedule json}
}
\value{
a data frame
}
\description{
parse_game() parses a single game from the Schedule >> Date JSON object and returns a data frame.
}
\note{
parse_game() is an inner function for \code{parse_date()}.
}
\seealso{
parse_date
}
