% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{ds.parse_highlight}
\alias{ds.parse_highlight}
\title{Parse Highlight}
\usage{
ds.parse_highlight(x)
}
\arguments{
\item{x}{the highlights json}
}
\value{
a data frame.
}
\description{
parse_highlight() parses a single highlight from the Highlights JSON object and returns a data frame
}
