% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{ds.parse_player}
\alias{ds.parse_player}
\title{Parse Player}
\usage{
ds.parse_player(x)
}
\arguments{
\item{x}{the pbp json}
}
\value{
data frame of player
}
\description{
parse_player() parses a single player from the PBP JSON object and returns a data frame.
}
