% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{ds.parse_roster}
\alias{ds.parse_roster}
\title{Parse Roster}
\usage{
ds.parse_roster(roster)
}
\arguments{
\item{x}{a roster object from ds.get_roster()}
}
\value{
a data frame of the rosters
}
\description{
parse_roster parses a roster object from ds.get_roster() and returns a data frame.
}
\keyword{internal}
