% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{ds.parse_shift}
\alias{ds.parse_shift}
\title{Parse Shift}
\usage{
ds.parse_shift(x)
}
\arguments{
\item{x}{a shift object}
}
\value{
a data frame of the shift
}
\description{
parse_shift() parses a single shift from the Shifts JSON object and returns a data frame.
}
\keyword{internal}
