% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{ds.parse_shifts}
\alias{ds.parse_shifts}
\title{Parse Shifts}
\usage{
ds.parse_shifts(player, venue, inner, outer)
}
\arguments{
\item{player}{player to parse}

\item{venue}{home or away}

\item{inner}{inner}

\item{outer}{outer}
}
\value{
matrix
}
\description{
parse_shifts() returns a matrix containing shift information for a single player
}
