% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{ds.parse_shifts_from_pbp}
\alias{ds.parse_shifts_from_pbp}
\title{Parse Shifts from PBP}
\usage{
ds.parse_shifts_from_pbp(shift_pbp_df, roster)
}
\arguments{
\item{shift_pbp_df}{the formatted shift pbp dataframe}

\item{roster}{a roster collected from the .htm}
}
\value{
shifts in a dataframe
}
\description{
Sometimes the .json or .htm shift reports are unavailable. This creates approximate shift table from the pbp.
}
\keyword{internal}
