% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{ds.scrape_game}
\alias{ds.scrape_game}
\title{Scrape Game}
\usage{
ds.scrape_game(season, game_id, try_tolerance = 3,
  agents = hockeyR::ds.get_user_agents()[sample(1:4, size = 1)])
}
\arguments{
\item{season}{The season of the game}

\item{game_id}{the game id to scrape and process}

\item{try_tolerance}{the number of tries before moving on.}

\item{agents}{a useragent, selected from the list of useragents.}
}
\value{
list object containing c([[1]] = PBP, [[2]] = Shifts, [[3]] = Highlights) is returned
}
\description{
scrape_game() collects and parses the data for a game corresponsing to a given season and game ID.
}
