% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{ds.scrape_player_profile}
\alias{ds.scrape_player_profile}
\title{Scrape Player Profile}
\usage{
ds.scrape_player_profile(player_id, try_tolerance = 3,
  agents = hockeyR::ds.get_user_agents()[sample(1:4, size = 1)])
}
\arguments{
\item{player_id}{the player ID to get}

\item{try_tolerance}{the number of tries before moving on.}

\item{agents}{a useragent, selected from the list of useragents.}
}
\value{
data frame of player info
}
\description{
scrape_player_profile() collects and parses the data for a player corresponsing to a given player ID.
}
