% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{ds.scrape_schedule}
\alias{ds.scrape_schedule}
\title{Scrape Schedule}
\usage{
ds.scrape_schedule(start, end, try_tolerance = 3,
  agents = hockeyR::ds.get_user_agents()[sample(1:4, size = 1)])
}
\arguments{
\item{start}{the start of the date range for the schedule}

\item{end}{the end of the date range for the schedule}

\item{try_tolerance}{the number of tries before moving on.}

\item{agents}{a useragent, selected from the list of useragents.}
}
\value{
data frame schedule
}
\description{
scrape_schedule() collects and parses the schedule data for a range corresponsing to a given start and end date.
}
