% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{ds.scrape_team_profile}
\alias{ds.scrape_team_profile}
\title{Scrape Team Profile}
\usage{
ds.scrape_team_profile(team_id, try_tolerance = 3,
  agents = ds.user_agents[sample(1:4, size = 1)])
}
\arguments{
\item{team_id}{Team ID to get}

\item{try_tolerance}{the number of tries before moving on.}

\item{agents}{a useragent, selected from the list of useragents.}
}
\value{
a data frame of team info
}
\description{
scrape_team_profile() collects and parses the data for a team corresponsing to a given team ID.
}
