% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{ds.seconds_from_ms}
\alias{ds.seconds_from_ms}
\title{Seconds from MS}
\usage{
ds.seconds_from_ms(ms)
}
\arguments{
\item{ms}{minute:second time}
}
\value{
time in seconds
}
\description{
seconds_from_ms() returns a numeric vector of representation in seconds of a given vector in M:S format.
}
\keyword{internal}
