% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{ds.who}
\alias{ds.who}
\title{Who}
\usage{
ds.who(player_id)
}
\arguments{
\item{player_id}{player id to get the name of}
}
\value{
the player's name
}
\description{
who() searches a given player ID and returns the player's full name
}
