% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_player_columns.R
\name{fix_player_columns}
\alias{fix_player_columns}
\title{Fix event player columns}
\usage{
fix_player_columns(players, col)
}
\arguments{
\item{players}{A tibble of event players from JSON pbp data}

\item{col}{Integer between 1 & 4 for which event player is being evaluated}
}
\value{
A tibble of an individual event player with cleaned column names
}
\description{
Helper function to parse player columns from JSON pbp data &
clean up the column names
}
