% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{ftable2df}
\alias{ftable2df}
\title{F Table to Data Frame}
\usage{
ftable2df(mydata)
}
\arguments{
\item{mydata}{an ftable object to be coerced to a data.frame}
}
\value{
a data frame coerced from mydata
}
\description{
ftable2df() returns a data.frame from an ftable object
}
