% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{get.st.game_score_weight}
\alias{get.st.game_score_weight}
\title{Get Game Score Weights}
\usage{
get.st.game_score_weight()
}
\value{
game score contribution weight.
}
\description{
Get the Game Score contribution weights
}
\note{
Game score vectors must be sorted in the same order as the weights. That is: `paste(names(get.st.game_score_weight()), collapse = ', ')`
}
