% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_rosters.R
\name{get_current_rosters}
\alias{get_current_rosters}
\title{Get current NHL rosters}
\usage{
get_current_rosters()
}
\value{
A tibble containing the current rosters for every team per the NHL
\describe{
\item{player_id}{NHL player ID, as an integer}
\item{player}{Player name as a string}
\item{jersey_number}{Player jersey number, as an integer}
\item{position}{Abbreviated official player position, as a string}
\item{position_type}{Abbreviated position group: F, D, or G}
\item{team_id}{NHL integer team ID}
\item{full_team_name}{Full team name as a string}
\item{team_abbr}{Team abbreviation, as a string}
}
}
\description{
Get current NHL rosters
}
\examples{
\dontrun{
current_rosters <- get_current_rosters()
}
}
