% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_game_info.R
\name{get_game_info}
\alias{get_game_info}
\title{Gather basic game information}
\usage{
get_game_info(game_id)
}
\arguments{
\item{game_id}{Game ID to scrape (Can be found using get_game_ids function)}
}
\value{
A 1xN tibble containing N pieces of information about the specified game
}
\description{
Scrapes basic game info like date, venue, & information about
the home and away teams for a given game
}
