% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_game_shifts.R
\name{get_game_shifts}
\alias{get_game_shifts}
\title{Fetch game shift data}
\usage{
get_game_shifts(game_id)
}
\arguments{
\item{game_id}{Game ID to scrape}
}
\value{
A tibble containing each player change as an event
}
\description{
A function to gather shift data from a given game. Shifts are
turned into events to match the style of events in standard game pbp.\cr\cr
Portions of this code are modified versions of code from the NHL scraper by
Evolving-Hockey, which in turn were modified from the NHL scraper by Manny Perry.
}
\examples{
\dontrun{
get_game_shifts(2020020561)
}
}
