% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_goalie_stats_hr.R
\name{get_goalie_stats_hr}
\alias{get_goalie_stats_hr}
\title{Scrape goalie stats}
\usage{
get_goalie_stats_hr(season = as.numeric(format(Sys.Date() + 81, "\%Y")))
}
\arguments{
\item{season}{Integer value denoting the end year of the season to scrape}
}
\value{
A tibble containing all goalie stats found on hockey-reference.com
for the given season
}
\description{
A function to scrape all goalie stats from a single season via
hockey-reference.com
}
\examples{
\dontrun{
get_goalie_stats_hr(2022)
}
}
