% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jersey_players.R
\name{get_jersey_players}
\alias{get_jersey_players}
\title{Get Players by Jersey Number}
\usage{
get_jersey_players(jersey)
}
\arguments{
\item{jersey}{An integer or a vector of integers between 0 & 99}
}
\value{
A tibble containing each player-season where a player wore
the specified number
}
\description{
Get the name of every player to wear a specific jersey number in
the NHL and the season in which they wore it.
}
\examples{
\dontrun{
#' get_jersey_players(69)
get_jersey_players(c(99, 66))
}
}
