% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_stats_hr.R
\name{get_player_stats_hr}
\alias{get_player_stats_hr}
\title{Get player counting stats from hockey-reference.com}
\usage{
get_player_stats_hr(player_name, season = "career", league = "NHL")
}
\arguments{
\item{player_name}{A player name or vector of player names}

\item{season}{An integer value denoting the end year of the season(s) to scrape}

\item{league}{The league stats to scrape, either 'NHL' or 'WHA'}
}
\value{
A tibble containing goals, assists, and various other stats for
the specified player(s) from hockey-reference.com
}
\description{
Get player counting stats from hockey-reference.com
}
\examples{
\dontrun{
#' get_player_stats("Wayne Gretzky")
get_player_stats_hr(c("Wayne Gretzky","Mario Lemieux"))
}
}
