% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rosters.R
\name{get_rosters}
\alias{get_rosters}
\title{Get team rosters from hockey-reference.com}
\usage{
get_rosters(team = "all", season = as.numeric(format(Sys.Date() + 184, "\%Y")))
}
\arguments{
\item{team}{A character vector of team names or abbreviations}

\item{season}{An integer value denoting the end year of the season to scrape}
}
\value{
A tibble containing the latest roster for the specified team(s)
in the specified season
}
\description{
Get team rosters from hockey-reference.com
}
\examples{
\dontrun{
get_rosters("SEA", 2022)
}
}
