% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_standings.R
\name{get_standings}
\alias{get_standings}
\title{Load team standings}
\usage{
get_standings(seasons = as.numeric(format(Sys.Date() + 78, "\%Y")))
}
\arguments{
\item{seasons}{End year of seasons to pull}
}
\value{
A tibble containing team records and stats for given year
}
\description{
Get full regular season standings for given year(s), including win-loss record and
goals for and against
}
\examples{
\dontrun{
get_standings(2022)
}
}
