% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_team_roster.R
\name{get_team_rosters}
\alias{get_team_rosters}
\title{Get current team roster}
\usage{
get_team_rosters(team)
}
\arguments{
\item{team}{A team name or abbreviation, as a string - or the NHL integer team ID}
}
\value{
A tibble containing the current official team roster per NHL.com
}
\description{
Get current team roster
}
\examples{
\dontrun{
get_team_rosters("SEA")
}
}
