% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{log_loss}
\alias{log_loss}
\title{Logarithmic Loss}
\usage{
log_loss(act, pred, allow_inf = FALSE)
}
\arguments{
\item{act}{actual result(s)}

\item{pred}{predicted result(s)}

\item{allow_inf}{whether to allow infinite penalties}
}
\value{
the log loss score for this data or series of data
}
\description{
log_loss() returns the logarithmic loss obtained from a given prediction and known result. The allow_inf parameter controls whether infinite loss is allowed (default is FALSE). Setting allow_inf to FALSE will cause large but finite penalties at the extremes.
}
