% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{moving}
\alias{moving}
\title{Moving Average}
\usage{
moving(x, n = 5)
}
\arguments{
\item{x}{a numerical vector to calculate the moving average}

\item{n}{the number of elements to include in the moving average}
}
\value{
the vector of the moving average.
}
\description{
moving() returns a vector of averages obtained from the n elements of x preceding and including the element at each respective index
}
