% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{na_as_string}
\alias{na_as_string}
\title{NA as String}
\usage{
na_as_string(x)
}
\arguments{
\item{x}{a character vector that may contain NA values to coerce to string "NA"}
}
\value{
a character vector with NA coerced to string "NA"
}
\description{
na_as_string() returns a character vector with NA values replaced as "NA"
}
