% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{na_as_zero}
\alias{na_as_zero}
\title{NA as Zero}
\usage{
na_as_zero(x)
}
\arguments{
\item{x}{a numeric vector that may contain NA values to coerce to 0}
}
\value{
a vector with NA values coerced to 0
}
\description{
na_as_zero() returns a numeric vector with NA values replaced as 0
}
