% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{na_if_null}
\alias{na_if_null}
\title{NA if NULL}
\usage{
na_if_null(x)
}
\arguments{
\item{x}{the object to try return}
}
\value{
the object's value, or NA
}
\description{
na_if_null() returns an object's value if it is not NULL and NA otherwise
}
