% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{nabs}
\alias{nabs}
\title{Numeric Absolute}
\usage{
nabs(x)
}
\arguments{
\item{x}{the value to coerce to numeric via character.}
}
\value{
a numeric value from x
}
\description{
nabs() returns x after first converting it to class numeric via character. Its primary use is converting objects of class factor to numeric. It also provides a more concise wrapper for standard numeric conversion.
}
