% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-not_in.R
\name{\%not_in\%}
\alias{\%not_in\%}
\title{'\%not_in\%' operator}
\usage{
lhs \%not_in\% rhs
}
\arguments{
\item{lhs}{A vector of values}

\item{rhs}{A vector of values to compare to lhs}
}
\value{
Returns \code{TRUE} if \code{lhs} is \strong{not} present in \code{rhs}
}
\description{
'\%not_in\%' operator
}
\keyword{internal}
