% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_nhl_prepare_xg_data.R
\name{helper_nhl_prepare_xg_data}
\alias{helper_nhl_prepare_xg_data}
\title{Prepare xG data}
\usage{
helper_nhl_prepare_xg_data(x)
}
\arguments{
\item{x}{A play-by-play data frame generated by hockeyR before xG is calculated}
}
\value{
A tibble; pbp data with xG model mutations along with identifiers for game and strength state
}
\description{
Helper function to prepare hockeyR pbp data for xG calculations
}
\examples{
\dontrun{
pbp <- load_pbp(2022) \%>\% dplyr::select(-xg)
model_data <- helper_nhl_prepare_xg_data(pbp)
}
}
\keyword{internal}
