% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_day.R
\name{scrape_day}
\alias{scrape_day}
\title{Scrape play-by-play for single day's games}
\usage{
scrape_day(day = as.Date(Sys.Date(), "\%Y-\%m-\%d"))
}
\arguments{
\item{day}{A day in the format of 'YYYY-MM-DD'; defaults to system date.}
}
\value{
A tibble containing all play-by-play data for a given day in
the same format as the output of \code{\link{scrape_game}}
}
\description{
Scrape play-by-play for single day's games
}
\examples{
\dontrun{
pbp_day <- scrape_day(day = "2015-01-06")
}
}
