% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_season.R
\name{scrape_season}
\alias{scrape_season}
\title{Scrape full season play-by-play}
\usage{
scrape_season(season, type = "all")
}
\arguments{
\item{season}{An integer value denoting the end year of the season to scrape}

\item{type}{A character vector of the game types to include: REG, POST, or "all"}
}
\value{
A tibble containing all play-by-play data for a given season in
the same format as the output of \code{\link{scrape_game}}
}
\description{
Scrape full season play-by-play
}
\examples{
\dontrun{
# scrape all regular season & postseason games for the 2016-2017 season
pbp_2016_2017 <- scrape_season(2017, type = "REG")
}
}
