% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{st.angle_from_centre}
\alias{st.angle_from_centre}
\title{Angle from Centre}
\usage{
st.angle_from_centre(x, y)
}
\arguments{
\item{x}{the x coordinate of the event}

\item{y}{the y coordinate of the event}
}
\value{
the angle from centre in degrees
}
\description{
# angle_from_centre() returns the angle from the central line perpendicular to the goal line in degrees of a location corresponsing to a given set of coordinates
}
