% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{st.combo_code}
\alias{st.combo_code}
\title{Combo Code}
\usage{
st.combo_code(p1, p2, p3)
}
\arguments{
\item{p1}{player 1 code}

\item{p2}{player 2 code}

\item{p3}{player 3 code}
}
\value{
a character string of the three player codes sorted with '-' separator
}
\description{
combo_code() returns the unique code produced from a list of up to three players.
}
