% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{st.distance_from_net}
\alias{st.distance_from_net}
\title{Distance from Net}
\usage{
st.distance_from_net(x, y)
}
\arguments{
\item{x}{the x coordinate of the event}

\item{y}{the y coordinate of the event}
}
\value{
the distance from the nearest net in feet
}
\description{
distance_from_net() returns the distance from the nearest net in feet of a location corresponding to a given set of coordinates
}
