% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{st.game_score}
\alias{st.game_score}
\title{Game Score}
\usage{
st.game_score(x)
}
\arguments{
\item{x}{the game score vector}
}
\value{
game score for that vector
}
\description{
game_score() returns the game score obtained from a given vector of statistics. The vector x is expected to contain the necessary stats in proper order
}
\seealso{
\code{\link{get.st.game_score_weight}()}
}
