% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{st.old_sum_goalie}
\alias{st.old_sum_goalie}
\title{Summarize Goalie Stats (Old PBP Format)}
\usage{
st.old_sum_goalie(x, venue)
}
\arguments{
\item{x}{a Corsica 1.0 PBP data frame object}

\item{venue}{home or away}
}
\value{
goalie counting stats for the home or away goalie(s) in the supplied pbp frame
}
\description{
old_sum_goalie() summarizes all goalie counting stats from a Corsica 1.0 PBP data frame object. x is expected to be a grouped data frame with home_goalie or away_goalie as a grouping variable for venue = "home" and venue = "away" respectively.
}
