% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{st.old_sum_team}
\alias{st.old_sum_team}
\title{Summarize Team Stats (Old PBP Format)}
\usage{
st.old_sum_team(x, venue)
}
\arguments{
\item{x}{a Corsica 1.0 PBP data frame object}

\item{venue}{home or away}
}
\value{
player counting stats for the home or away players in the supplied pbp frame
}
\description{
old_sum_team() summarizes all team counting stats from a Corsica 1.0 PBP data frame object. x is expected to be a grouped data frame with home_team or away_team as a grouping variable for venue = "home" and venue = "away" respectively
}
