% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{st.pbp_enhance}
\alias{st.pbp_enhance}
\title{Enhance PBP}
\usage{
st.pbp_enhance(pbp)
}
\arguments{
\item{pbp}{a Corsica PBP data frame object}
}
\value{
an enahnced Corsica PBP data frame object (ePBP)
}
\description{
pbp_enhance() performs some preliminary operations on a given PBP data frame object and returns the enhanced version
}
