% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{st.sum_goalie}
\alias{st.sum_goalie}
\title{Summarize Goalie Stats}
\usage{
st.sum_goalie(x, venue)
}
\arguments{
\item{x}{a Corsica PBP data frame object}

\item{venue}{home or away}
}
\value{
player counting stats for the home or away goalie(s) in the supplied pbp frame
}
\description{
sum_goalie() summarizes all goalie counting stats from a PBP data frame object x is expected to be a grouped data frame with home_goalie or away_goalie as a grouping variable for venue = "home" and venue = "away" respectively
}
