% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{st.sum_skater}
\alias{st.sum_skater}
\title{Summarize Skater Stats}
\usage{
st.sum_skater(x, venue)
}
\arguments{
\item{x}{a Corsica PBP data frame object}

\item{venue}{home or away}
}
\value{
player counting stats for the home or away players in the supplied pbp frame
}
\description{
sum_skater() summarizes all skater counting stats from a PBP data frame object. x is expected to be a grouped data frame with home_on_x or away_on_x as a grouping variable for venue = "home" and venue = "away" respectively.
}
\note{
A dplyr::rename() argument must be passed before sum_skater() to convert home/away_on_x to player
}
