% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{st.sum_team}
\alias{st.sum_team}
\title{Summarize Team Stats}
\usage{
st.sum_team(x, venue)
}
\arguments{
\item{x}{a Corsica PBP data frame object}

\item{venue}{home or away}
}
\value{
player counting stats for the home or away team in the supplied pbp frame
}
\description{
sum_team() summarizes all team counting stats from a ePBP data frame object. x is expected to be a grouped data frame with home_team or away_team as a grouping variable for venue = "home" and venue = "away" respectively
}
