% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{st.which_circle}
\alias{st.which_circle}
\title{Which Circle}
\usage{
st.which_circle(x, y)
}
\arguments{
\item{x}{The x coordinate of the event}

\item{y}{The y coordinate of the event}
}
\value{
the faceoff circle closest to the event
}
\description{
which_circle() returns the faceoff circle number nearest to a location corresponding to a given set of coordinates
}
