% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryscrape.R
\name{st.which_zone}
\alias{st.which_zone}
\title{Which Zone}
\usage{
st.which_zone(x)
}
\arguments{
\item{x}{the x coordinate of the event}
}
\value{
The zone as L, N, or R as a character
}
\description{
which_zone() returns the absolute zone of a location corresponding to a given x-coordinate
}
