% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_logos_colors_data.R
\docType{data}
\name{team_logos_colors}
\alias{team_logos_colors}
\title{Team logos & colors}
\format{
A data frame with 35 rows and 12 variables:
\describe{
\item{full_team_name}{full team name}
\item{team_abbr}{NHL.com team abbreviation}
\item{team_nick}{lowercase, no spaces team nickname}
\item{division}{current NHL division}
\item{conference}{current NHL conference}
\item{team_logo_espn}{primary team logo from ESPN.com}
\item{team_color1}{current primary team color}
\item{team_color2}{current secondary team color}
\item{team_logo_alternate}{alternate or throwback logo}
\item{team_color_alt1}{alternate logo primary color}
\item{team_color_alt2}{alternate logo secondary color}
\item{status}{active or inactive}
}
}
\source{
\url{https://www.espn.com/nhl/teams}

\url{https://www.sportslogos.net/teams/list_by_league/1/National_Hockey_League/NHL/logos/}

\url{https://teamcolorcodes.com/nhl-team-color-codes/}
}
\usage{
team_logos_colors
}
\description{
A dataset containing the full team names, abbreviations, colors & logos
for all 32 NHL teams (plus some defunct teams)
}
\keyword{datasets}
