% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xGoals.R
\name{xg.train_model}
\alias{xg.train_model}
\title{Train xG model}
\usage{
xg.train_model(data_dir = "./data", train_season = c(2016:2017),
  test_season = c(2018))
}
\arguments{
\item{data_dir}{the data directory with compiled .pbp files in season subfolders, also where to save model}

\item{train_season}{vector of seasons (as season end year as integer) to use for training}

\item{test_season}{vector of seasons (as season end year as integer) to use for testing}
}
\value{
the xGmodel
}
\description{
Train xG model
}
