% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_xg.R
\name{calculate_xg}
\alias{calculate_xg}
\title{Calculate hockeyR expected goals (xG)}
\usage{
calculate_xg(pbp)
}
\arguments{
\item{pbp}{A play-by-play data frame, previously returned by hockeyR::scrape_game}
}
\value{
The original supplied play-by-play data with a column for expected goals appended
}
\description{
Uses the hockeyR expected goals model to calculate xG for any pbp data frame generated by hockeyR
}
\examples{
\dontrun{
pbp <- load_pbp(2022) \%>\% dplyr::select(-xg)
pbp_preds <- calculate_xg(pbp)
}
}
