% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_game_info.R
\name{get_game_info}
\alias{get_game_info}
\title{Gather basic game information}
\usage{
get_game_info(site)
}
\arguments{
\item{site}{Raw JSON data from NHL API}
}
\value{
A 1xN tibble containing N pieces of information about the specified game
}
\description{
A helper function used in game scraping. Adds columns like date,
venue, & information about the home and away teams
}
