% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hockeyR-package.R
\docType{package}
\name{hockeyR-package}
\alias{hockeyR}
\alias{hockeyR-package}
\title{hockeyR: Collect and Clean Hockey Stats}
\description{
Various functions to scrape hockey play-by-play data from the
\url{https://www.nhl.com/}. It also contains functions to scrape data from \url{https://www.hockey-reference.com/},
including standings, player stats, and jersey number history.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/danmorse314/hockeyR}
  \item Report bugs at \url{https://github.com/danmorse314/hockeyR/issues}
}

}
\author{
\strong{Maintainer}: Daniel Morse \email{danmorse8642@gmail.com}

}
\keyword{internal}
