% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-not_in.R
\name{\%not_in\%}
\alias{\%not_in\%}
\title{'\%not_in\%' operator}
\usage{
x \%not_in\% table
}
\arguments{
\item{x}{A vector of values}

\item{table}{A vector of values to compare to \code{table}}
}
\value{
Returns \code{TRUE} if \code{x} is \strong{not} present in \code{table}
}
\description{
'\%not_in\%' operator
}
\keyword{internal}
